Refactored correctly and added simplicity with proper functionality at the same time.
