+++ app/repositories/guides.py
